move EEG referencing code to separate file
